refactor: components in component folder
