+++ frontend/src/app/core/components/footer/footer.component.html
@@ -1,4 +1,4 @@
-<footer class="pt-5 text-muted text-center text-small fixed-bottom">
+<footer class="pt-5 text-muted text-center text-small">
   <p class="mb-1">{{year}} Made by Cássio Fernando</p>
   <ul class="list-inline">
     <li class="list-inline-item"><a href="https://github.com/cassiofb-dev" target="_blank" rel="noopener">Github</a></li>
