+++ frontend/src/app/core/components/footer/footer.component.ts
