+++ frontend/src/app/core/components/footer/footer.module.ts
