+++ frontend/src/app/core/components/seo/seo.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Meta, Title } from '@angular/platform-browser';
-import { ISEO } from '../interfaces/seo.inteface';
+import { ISEO } from '../../interfaces/seo.inteface';
 
 @Component({
   selector: 'app-seo',
